Revert "environment variables created"
