+++ .gitignore
@@ -1,3 +1,2 @@
 server/node_modules
 server/package-lock.json
-server/.env
