+++ README.md
@@ -24,4 +24,3 @@ Link to [live version](https://3daddict.github.io/oop-todo/).
   * Refreshed understanding of Bootstrap 4 documentation and new features.
   * Better utilization of Sass.
   * OOP Components and working with constructor functions
-  * Setting up NodeJS Back-End with Routes, End Points and Environment Variables
