--- server/config/db_creds.js
@@ -1,11 +1,0 @@
-require('dotenv').config();
-const mysql = require('mysql');
-
-const con = mysql.createConnection({
-    host: process.env.DB_HOST,
-    user: process.env.DB_USER,
-    password: process.env.DB_PASSWORD,
-    database: process.env.DB_DATABASE
-});
-
-module.exports = con;
