+++ server/package.json
@@ -5,7 +5,6 @@
   "main": "server.js",
   "dependencies": {
     "body-parser": "^1.18.3",
-    "dotenv": "^6.2.0",
     "express": "^4.16.4",
     "morgan": "^1.9.1",
     "mysql": "^2.16.0",
